Split bin/ into bin/ (operator scripts) and scripts/ (build helpers) (#287)
Move npm/webpack-invoked tooling helpers out of bin/ into a new scripts/
directory, leaving bin/ for operator-run scripts (AWS deploy/download/
invalidate, Docker server launchers).

Moved to scripts/: build-datasets-manifest.js (+ __tests__), setup-eslint-plugin.js,
postbuild.sh. Updated call sites in package.json (setup-eslint-plugin),
server.js (require + collision-warning command string), and
webpack.config.prod.js (postbuild.sh path), plus docs (CLAUDE.md,
DEVELOPMENT.md, DEPLOYMENT_SETUP.md) and the CHANGELOG.

Verified: require resolves, npm run setup-eslint-plugin runs, full Jest suite
662/662 passes, npm run build succeeds with postbuild.sh running from scripts/.
