+++ scripts/setup-eslint-plugin.js
